requirements.txt
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
Flask==2.3.3
Flask-Cors==4.0.0
flask-ngrok==0.0.25
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
outcome==1.2.0
Pillow==10.0.1
pycparser==2.21
pyngrok==7.0.0
pypng==0.20220715.0
PySocks==1.7.1
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==2.3.7
wsproto==1.2.0
